requirements.txt
aiohttp==3.9.5
websockets==12.0
pydantic==2.8.2
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
cryptography==43.0.1
base58==2.1.1
solana==0.30.2
solders==0.21.0
